model/requirements.txt
fastapi
uvicorn
transformers
torch
faiss-cpu
scikit-learn
pydantic
numpy
pickle-mixin
